+++ ueueo-framework/ueueo-multi-tenancy/src/main/java/com/ueueo/multitenancy/ITenantStore.java
@@ -8,6 +8,6 @@ public interface ITenantStore {
 
     TenantConfiguration find(String name);
 
-    TenantConfiguration find(Integer id);
+    TenantConfiguration find(Long id);
 
 }
